// src/app/shared/services/claim/entities/claim-by-member.dto.ts
/**
* ClaimDto.ts
*
* @description: DTO claims
* @author Juan Camilo Moreno.
* @version 1.0
* @date 10-10-2018.
*
**/

import { Claim } from './claims';

/**
 * DTO claims
 */
export interface ClaimByMemberDto {

    /**
     * Count
     */
    totalCount: number;

    /**
     * pageIndex
     */
    pageIndex: number;

    /**
     * pageSize
     */
    pageSize: number;

    /**
     * data: Claim[]
     */
    list: Claim[];
}